Default under_six to empty and require it when saving
These two changes support product functionality, where we want the
user to explicitly choose an age group before allowing them to
continue with the assessment. By defaulting to null on the
prototype, we allow for an 'unchosen' state, and by requiring it
before saving, we ensure that they did pick one.